Move Misc items into components folder
